chore(map): 🚚 | move egg and player management to more appropriate files
