+++ Makefile
@@ -74,7 +74,8 @@ SRC_SERVER = $(SRCDIR_SERVER)main.c	\
 			 $(SRCDIR_SERVER)options_parser/options/debug.c	\
 			 $(SRCDIR_SERVER)map/map.c	\
 			 $(SRCDIR_SERVER)map/coordinates.c	\
-			 $(SRCDIR_SERVER)map/add_and_remove.c	\
+			 $(SRCDIR_SERVER)map/player_management.c	\
+			 $(SRCDIR_SERVER)map/egg_management.c	\
 			 $(SRCDIR_SERVER)player/player.c	\
 			 $(SRCDIR_SERVER)player/movement.c	\
 			 $(SRCDIR_SERVER)map/resources.c	\
@@ -177,7 +178,8 @@ TESTS_SRC =	$(SRCDIR_SERVER)map/map.c	\
 			${SRCDIR_SERVER}map/resources.c	\
 			$(SRCDIR_SERVER)player/player.c	\
 			$(SRCDIR_SERVER)player/movement.c	\
-			$(SRCDIR_SERVER)map/add_and_remove.c	\
+			$(SRCDIR_SERVER)map/player_management.c	\
+			$(SRCDIR_SERVER)map/egg_management.c	\
 			$(SRCDIR_SERVER)egg/egg.c	\
 			${TESTS}player_tests.c	\
 			${TESTS}resources_tests.c	\
